remove unused css rules and add splitChunks to webpack
