+++ src/common.js
@@ -3,3 +3,4 @@ import './images/tech-radar-landing-page-wide.png';
 import './images/tw-logo.png';
 import './images/favicon.ico';
 import './images/search-logo-2x.svg';
+import 'es6-promise/auto';
