+++ src/graphing/graphing.js
@@ -454,7 +454,7 @@ export default class Graphing {
 
         tmpHeader.append('br');
 
-        this.header.append('div')
+        tmpHeader.append('div')
             .attr('class', 'headerpic')
             .html('<a href="/" target="_top"><img class="img-fluid" src="images/headercomp.png" alt="Logo"/></a>');
